requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.5
django-admin-relation-links==0.2.4
django-heroku==0.3.1
django-telegrambot==1.0.1
django-timezone-field==3.1
future==0.17.1
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
oauth2client==4.1.3
psycopg2==2.8.3
pyaes==1.6.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
Pyrogram==0.16.0
PySocks==1.7.0
python-telegram-bot==12.1.1
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.6
whitenoise==4.1.4
django-dbconn-retry
